Move `with_tracing_logs` to chalk-solve
