+++ chalk-solve/Cargo.toml
@@ -14,6 +14,7 @@ ena = "0.14.0"
 itertools = "0.9.0"
 petgraph = "0.5.0"
 tracing = "0.1"
+tracing-subscriber = "0.2"
 rustc-hash = { version = "1.0.0" }
 
 chalk-derive = { version = "0.14.0-dev.0", path = "../chalk-derive" }
