+++ tests/test/mod.rs
@@ -5,10 +5,11 @@ use chalk_integration::interner::ChalkIr;
 use chalk_integration::lowering::LowerGoal;
 use chalk_integration::query::LoweringDatabase;
 use chalk_solve::ext::*;
+use chalk_solve::logging::with_tracing_logs;
 use chalk_solve::RustIrDatabase;
 use chalk_solve::{Solution, SolverChoice};
 
-use crate::test_util::{assert_same, with_tracing_logs};
+use crate::test_util::assert_same;
 
 #[cfg(feature = "bench")]
 mod bench;
